Refactor: 해시태그 추가부분 모듈화
