+++ src/components/organisms/index.ts
@@ -1,3 +1,4 @@
+export { default as AdminWriteForm } from './AdminWriteForm';
 export { default as CategoryList } from './CategoryList';
 export { default as ChallengeDetailCard } from './ChallengeDetailCard';
 export { default as ChallengeEditCard } from './ChallengeEditCard';
@@ -23,3 +24,4 @@ export { default as NoticeList } from './NoticeList';
 export { default as RegisterDetailCard } from './RegisterDetailCard';
 export { default as SearchForm } from './SearchForm';
 export { default as SetNewPwd } from './SetNewPwd';
+export { default as HashTagAddForm } from './HashTagAddForm';
